1-strncat.c

+++ 0x06-pointers_arrays_strings/1-strncat.c
@@ -9,8 +9,8 @@
  */
 char *_strncat(char *dest, char *src, int n)
 {
-	size_t des_len = strlen(dest);
-	size_t i;
+	int des_len = strlen(dest);
+	int i;
 
 	for (i = 0; i < n && src[i] != '\0'; i++)
 	dest[des_len + i] = src[i];
